3073: files(ets) over 2,147,483,647 bytes (2.14GB) will not be indexed (#515)
https://cits.artic.edu/issues/3073

+++ app/indexers/file_set_indexer.rb
@@ -5,7 +5,7 @@ class FileSetIndexer < CurationConcerns::FileSetIndexer
 
   def generate_solr_document
     super.tap do |solr_doc|
-      solr_doc[Solrizer.solr_name("file_size", :stored_sortable, type: :integer)] = object.file_size[0]
+      solr_doc[Solrizer.solr_name("file_size", :stored_sortable, type: :long)] = object.file_size[0]
       solr_doc[Solrizer.solr_name("image_height", :searchable, type: :integer)] = dimensions.height
       solr_doc[Solrizer.solr_name("image_width", :searchable, type: :integer)] = dimensions.width
       solr_doc[Solrizer.solr_name("rdf_types", :symbol)] = object.type.map(&:to_s)
